Rewrite versions_tests in pytest form.
